bug when get_all_app_list has no app fixed

+++ sproutime.py
@@ -884,7 +884,11 @@ class Tabview(ctk.CTkTabview):
         tab_list = [tab for tab in constant_tab_list if tab not in quest_list]
         self.tabBox.configure(values=tab_list)
         
-        temp_quest_app = app_list[0]
+        if app_list:
+            temp_quest_app = app_list[0]
+        else:
+            temp_quest_app = ""
+
         temp_quest_tab = tab_list[0]
         
         self.check_for_chrome()
@@ -967,39 +971,40 @@ class Tabview(ctk.CTkTabview):
         
     def save_quest_time(self):
         global temp_quest_app, temp_quest_tab, quest_list_update, quest_complete_update, app_time_update, treeview_update
-        minutes = slider_var.get() * 60
-        name = temp_quest_tab if temp_quest_app == google and temp_quest_tab != "Any Tabs" else temp_quest_app
-        if switch_var.get() == ">":
-            maximum = 1
-        else:
-            maximum = 0
-            completed_list.append(name)
+        if temp_quest_app:
+            minutes = slider_var.get() * 60
+            name = temp_quest_tab if temp_quest_app == google and temp_quest_tab != "Any Tabs" else temp_quest_app
+            if switch_var.get() == ">":
+                maximum = 1
+            else:
+                maximum = 0
+                completed_list.append(name)
 
-        conn = get_database_connection(APP_NAME, DB_RELATIVE_PATH, DATABASE_FILENAME)
-        cursor = conn.cursor()
-        
-        try:
-            cursor.execute("INSERT INTO quest (app_name, time, maximum) VALUES (?, ?, ?)", (name, minutes, maximum))
-            conn.commit()
+            conn = get_database_connection(APP_NAME, DB_RELATIVE_PATH, DATABASE_FILENAME)
+            cursor = conn.cursor()
             
-            quest_list.append(name)
-            quest_dict[name] = {"maximum": switch_var.get(), "time": minutes * 60}
-            
-            #Check if quest is completed today
-            cursor.execute("DELETE FROM activity_log WHERE date = ? AND app_name = ? AND type = 1", (str(date.today()), name))
-            conn.commit()
+            try:
+                cursor.execute("INSERT INTO quest (app_name, time, maximum) VALUES (?, ?, ?)", (name, minutes, maximum))
+                conn.commit()
+                
+                quest_list.append(name)
+                quest_dict[name] = {"maximum": switch_var.get(), "time": minutes * 60}
+                
+                #Check if quest is completed today
+                cursor.execute("DELETE FROM activity_log WHERE date = ? AND app_name = ? AND type = 1", (str(date.today()), name))
+                conn.commit()
 
-        except sqlite3.Error as e:
-            if DEBUG: print(f"An SQL error occurred: {e}")
-            conn.rollback()
-        finally:
-            if conn:
-                conn.close()
-            
-        check_quest(app_name)
-        self.refresh_app_list()
+            except sqlite3.Error as e:
+                if DEBUG: print(f"An SQL error occurred: {e}")
+                conn.rollback()
+            finally:
+                if conn:
+                    conn.close()
+                
+            check_quest(app_name)
+            self.refresh_app_list()
 
-        quest_list_update = True
+            quest_list_update = True
         
     def refresh_stat(self):
         global stat_update
@@ -1683,7 +1688,7 @@ def get_all_app_list():
 
                 if foundPID == pid and win32gui.IsWindowVisible(handle):
                     process_name = process.info["name"]
-                    if process_name not in ignored_processes:
+                    if (process_name not in ignored_processes) and (process_name != "Sproutime.exe"):
                         app_name = process_name.split(".")[0].capitalize() # Get all active app name
                         if app_name not in app_list:
                             app_list.append(app_name)
@@ -2108,7 +2113,7 @@ completed_list = []
 failed_list = []
 slider_var = ctk.IntVar(value=1)
 switch_var = ctk.StringVar(value=">")
-ignored_processes = ["","Sproutime.exe", "explorer.exe", "TextInputHost.exe", "ApplicationFrameHost.exe", "Taskmgr.exe", "SearchHost.exe", "ShellExperienceHost.exe"]
+ignored_processes = ["", "explorer.exe", "TextInputHost.exe", "ApplicationFrameHost.exe", "Taskmgr.exe", "SearchHost.exe", "ShellExperienceHost.exe"]
 
 # Progress Tab UI
 progressbar_dict = {}
@@ -2156,7 +2161,10 @@ theme_options = ["Dark", "Light", "System"]
 allow_noti = True
 theme = "Dark"
 secondary_colour = "#515151"
-temp_quest_app = app_list[0]
+if app_list:
+    temp_quest_app = app_list[0]
+else:
+    temp_quest_app = ""
 temp_quest_tab = tab_list[0]
 
 p1 = threading.Thread(target=update_time)
